Tweaks to simplify DB setup with H2 as default, support for Postgres in mind.
Squashed commit of the following:

commit 9312fb56d2eac3d78202aa41e3d61173ed0463e5
Author: Matt Magoffin <matt@solarnetwork.net>
Date:   Fri Jun 6 10:47:26 2025 +1200

    More tweaks for H2 as default.

commit 307e0949ae0a9ffb9fd1a79e5881a6169c965f4e
Author: Matt Magoffin <matt@solarnetwork.net>
Date:   Fri Jun 6 10:37:51 2025 +1200

    More tweaks for H2 as default.

commit 7638ebe690609bb60e6c01efc8e1110831fb6567
Author: Matt Magoffin <matt@solarnetwork.net>
Date:   Fri Jun 6 07:54:32 2025 +1200

    Switch to H2 as default test database.

commit 9e961ad91b7fbd17ff989c0beb5e575d73beb524
Author: Matt Magoffin <matt@solarnetwork.net>
Date:   Fri Jun 6 07:53:35 2025 +1200

    Drop derby- default prefix.

+++ net.solarnetwork.node.dao.jdbc.test/src/net/solarnetwork/node/dao/jdbc/test/csv-data-01.sql
@@ -2,7 +2,7 @@ CREATE TABLE solarnode.test_csv_io (
 	pk	  BIGINT NOT NULL,
 	str	  VARCHAR(255),
 	inum  INTEGER,
-	dnum  DOUBLE,
+	dnum  DOUBLE PRECISION,
 	ts    TIMESTAMP,
 	PRIMARY KEY (pk)
 );
